# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CallSummary": {
        "properties": {
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/CallChange"
            },
            "title": "Appointments",
            "type": "array"
          },
          "clinic_id": {
            "title": "Clinic Id",
            "type": "integer"
          },
          "clinic_name": {
            "title": "Clinic Name",
            "type": "string"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "end_reason": {
            "title": "End Reason",
            "type": "string"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "stack": {
            "title": "Stack",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "live",
              "ended",
              "dropped"
            ],
            "title": "Status",
            "type": "string"
          },
          "tool_failures": {
            "title": "Tool Failures",
            "type": "integer"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "clinic_id",
          "clinic_name",
          "started_at",
          "duration_s",
          "status",
          "end_reason",
          "outcome",
          "stack",
          "turn_count",
          "error_count",
          "tool_failures",
          "appointments"
        ],
        "title": "CallSummary",
        "type": "object"
      },
      "CallTrace": {
        "properties": {
          "accesses": {
            "items": {
              "$ref": "#/components/schemas/TranscriptAccess"
            },
            "title": "Accesses",
            "type": "array"
          },
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/CallChange"
            },
            "title": "Appointments",
            "type": "array"
          },
          "clinic_id": {
            "title": "Clinic Id",
            "type": "integer"
          },
          "clinic_name": {
            "title": "Clinic Name",
            "type": "string"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "end_reason": {
            "title": "End Reason",
            "type": "string"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "stack": {
            "title": "Stack",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "live",
              "ended",
              "dropped"
            ],
            "title": "Status",
            "type": "string"
          },
          "tool_failures": {
            "title": "Tool Failures",
            "type": "integer"
          },
          "transcript_kept": {
            "title": "Transcript Kept",
            "type": "boolean"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "clinic_id",
          "clinic_name",
          "started_at",
          "duration_s",
          "status",
          "end_reason",
          "outcome",
          "stack",
          "turn_count",
          "error_count",
          "tool_failures",
          "appointments",
          "events",
          "transcript_kept",
          "accesses"
        ],
        "title": "CallTrace",
        "type": "object"
      },
      "CallsToday": {
        "description": "Today in the clinic's timezone: how busy the receptionist was, and the\ncalls where the caller may not have been helped (cut off or failed).",
        "properties": {
          "calls": {
            "title": "Calls",
            "type": "integer"
          },
          "changed": {
            "title": "Changed",
            "type": "integer"
          },
          "questions": {
            "title": "Questions",
            "type": "integer"
          },
          "unhelped": {
            "items": {
              "$ref": "#/components/schemas/ClinicCall"
            },
            "title": "Unhelped",
            "type": "array"
          }
        },
        "required": [
          "calls",
          "changed",
          "questions",
          "unhelped"
        ],
        "title": "CallsToday",
        "type": "object"
      },
      "Clinic": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "booking_window_days": {
            "title": "Booking Window Days",
            "type": "integer"
          },
          "call_link": {
            "title": "Call Link",
            "type": "string"
          },
          "doctors": {
            "items": {
              "$ref": "#/components/schemas/Doctor"
            },
            "title": "Doctors",
            "type": "array"
          },
          "faq": {
            "items": {
              "$ref": "#/components/schemas/Faq"
            },
            "title": "Faq",
            "type": "array"
          },
          "has_calls": {
            "title": "Has Calls",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "slots_offered": {
            "title": "Slots Offered",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "time_off": {
            "items": {
              "$ref": "#/components/schemas/TimeOff"
            },
            "title": "Time Off",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "call_link",
          "address",
          "phone",
          "timezone",
          "booking_window_days",
          "slots_offered",
          "doctors",
          "faq",
          "time_off",
          "has_calls"
        ],
        "title": "Clinic",
        "type": "object"
      },
      "ClinicCall": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ClinicCallChange"
            },
            "title": "Changes",
            "type": "array"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "live",
              "ended",
              "dropped"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "started_at",
          "duration_s",
          "status",
          "outcome",
          "changes"
        ],
        "title": "ClinicCall",
        "type": "object"
      },
...
      "ClinicDetails": {
        "properties": {
          "address": {
            "default": "",
            "maxLength": 500,
            "title": "Address",
            "type": "string"
          },
          "booking_window_days": {
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Booking Window Days",
            "type": "integer"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "default": "",
            "maxLength": 20,
            "title": "Phone",
            "type": "string"
          },
          "slots_offered": {
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Slots Offered",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "booking_window_days",
          "slots_offered"
        ],
        "title": "ClinicDetails",
        "type": "object"
      },
      "ClinicSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "staff": {
            "title": "Staff",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "staff"
        ],
        "title": "ClinicSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Member": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "Member",
        "type": "object"
      },
      "MemberIn": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "MemberIn",
        "type": "object"
      },
      "NewClinic": {
        "properties": {
          "address": {
            "default": "",
            "maxLength": 500,
            "title": "Address",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "default": "",
            "maxLength": 20,
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "NewClinic",
        "type": "object"
      },
...
      "SlugIn": {
        "properties": {
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug"
        ],
        "title": "SlugIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clinic Console API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/calls": {
      "get": {
        "operationId": "calls_api_admin_calls_get",
        "parameters": [
          {
            "in": "query",
            "name": "clinic_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clinic Id"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "with_errors",
            "required": false,
            "schema": {
              "default": false,
              "title": "With Errors",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CallSummary"
                  },
                  "title": "Response Calls Api Admin Calls Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calls"
      }
    },
    "/api/admin/calls/{call_id}": {
      "get": {
        "operationId": "call_trace_api_admin_calls__call_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "title": "Call Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallTrace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Call Trace"
      }
    },
...
    "/api/clinics": {
      "post": {
        "operationId": "create_clinic_api_clinics_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewClinic"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Clinic"
      }
    },
    "/api/clinics/{clinic_id}": {
      "get": {
        "operationId": "get_clinic_api_clinics__clinic_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Clinic"
      },
      "patch": {
        "operationId": "update_clinic_api_clinics__clinic_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Clinic"
      }
    },
...
    "/api/clinics/{clinic_id}/calls": {
      "get": {
        "operationId": "calls_api_clinics__clinic_id__calls_get",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClinicCall"
                  },
                  "title": "Response Calls Api Clinics  Clinic Id  Calls Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calls"
      }
    },
    "/api/clinics/{clinic_id}/calls/today": {
      "get": {
        "description": "Declared before /{call_id}, which would otherwise catch \"today\".",
        "operationId": "today_api_clinics__clinic_id__calls_today_get",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallsToday"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Today"
      }
    },
    "/api/clinics/{clinic_id}/calls/{call_id}": {
      "get": {
        "operationId": "call_api_clinics__clinic_id__calls__call_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "title": "Call Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicCallDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Call"
      }
    },
...
    "/api/clinics/{clinic_id}/members": {
      "get": {
        "operationId": "members_api_clinics__clinic_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  },
                  "title": "Response Members Api Clinics  Clinic Id  Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Members"
      },
      "post": {
        "operationId": "add_member_api_clinics__clinic_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Member"
      }
    },
    "/api/clinics/{clinic_id}/members/{member_id}": {
      "delete": {
        "operationId": "remove_member_api_clinics__clinic_id__members__member_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "title": "Member Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Remove Member Api Clinics  Clinic Id  Members  Member Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Member"
      }
    },
    "/api/clinics/{clinic_id}/slug": {
      "put": {
        "operationId": "set_slug_api_clinics__clinic_id__slug_put",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "title": "Clinic Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlugIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Slug"
      }
    },
...
  }
}